Project Complete
Invincibility PowerUp

+++ Assets/Scripts/Enemy/Killer.cs
@@ -7,6 +7,9 @@ public class Killer : Enemy
     protected override void PlayerImpact(Player player)
     {
         //base.PlayerImpact(player);
-        player.Kill();
+        if (player.CanBeDamaged)
+        {
+            player.Kill();
+        }
     }
 }
